ainet-voice-agent/pyproject.toml
[project]
name = "ainet-voice-agent"
version = "0.1.0"
description = "AiNet Voice Agent - LiveKit pipeline with Deepgram STT/TTS and OpenAI LLM"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=0.8",
    "livekit-plugins-deepgram>=0.6",
    "livekit-plugins-openai>=0.6",
    "livekit-plugins-silero>=0.6",
    "supabase>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.scripts]
start = "python -m agent.main"
